Link the flamegraphs in the benchmark results comment
